pyproject.toml
[tool.poetry]
name = "llama-server"
version = "0.1.0"
description = ""
authors = ["Alexis Craye <alexis.craye@rideyego.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
python-socketio = "^5.8.0"
aiohttp = "^3.8.4"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
torch = "^2.0.0"
numpy = "^1.24.2"
sentencepiece = "^0.1.97"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
